refactor(store): split repository/ into repository.rs, builder.rs, snapshot.rs
Promote three crate-root modules out of the repository/ subdirectory:

- repository.rs (~475 lines): Repository<A> trait + EventStore facade +
  ZeroCopyEventStore facade + pub(crate) ReplayFrom trait. The four files
  that were repository/{repository, event_store, zero_copy, replay}.rs.

- builder.rs (~390 lines): RepositoryBuilder typestate + NeedsCodec marker
  + NoSnapshot/WithSnapshot snapshot slot markers. Pure construction
  scaffolding, kept separate from the trait + impls.

- snapshot.rs (~150 lines): Snapshotting<R, SS, T> decorator
  (feature = "snapshot"). Wraps a Repository with transparent state-store
  load/save and a PersistTrigger policy.

Three files committed together because Rust's module resolution forbids
the intermediate state where repository.rs at root and repository/mod.rs
coexist — the move must be atomic.

Public crate-root paths unchanged via lib.rs re-exports:
- nexus_store::{Repository, EventStore, ZeroCopyEventStore}
- nexus_store::{RepositoryBuilder, NeedsCodec, NoSnapshot}
- nexus_store::{Snapshotting, WithSnapshot} (feature = "snapshot")

Sub-paths changed (no external consumer was using them):
- nexus_store::repository::RepositoryBuilder -> nexus_store::builder::RepositoryBuilder
- nexus_store::repository::Snapshotting     -> nexus_store::snapshot::Snapshotting
