+++ crates/nexus-store/src/lib.rs
@@ -1,9 +1,12 @@
+pub mod builder;
 pub mod codec;
 pub mod envelope;
 pub mod error;
 #[cfg(feature = "projection")]
 pub mod projection;
 pub mod repository;
+#[cfg(feature = "snapshot")]
+pub mod snapshot;
 pub mod state;
 pub mod store;
 #[cfg(feature = "testing")]
@@ -11,6 +14,9 @@ pub mod testing;
 pub mod upcasting;
 
 pub use arrayvec::ArrayString;
+#[cfg(feature = "snapshot")]
+pub use builder::WithSnapshot;
+pub use builder::{NeedsCodec, NoSnapshot, RepositoryBuilder};
 #[cfg(feature = "json")]
 pub use codec::serde::json::{Json, JsonCodec};
 #[cfg(feature = "serde")]
@@ -21,11 +27,9 @@ pub use error::{AppendError, InvalidSchemaVersion, StoreError, UpcastError};
 pub use nexus::Version;
 #[cfg(feature = "projection")]
 pub use projection::Projector;
-pub use repository::{
-    EventStore, NeedsCodec, NoSnapshot, Repository, RepositoryBuilder, ZeroCopyEventStore,
-};
+pub use repository::{EventStore, Repository, ZeroCopyEventStore};
 #[cfg(feature = "snapshot")]
-pub use repository::{Snapshotting, WithSnapshot};
+pub use snapshot::Snapshotting;
 #[cfg(feature = "testing")]
 pub use state::InMemoryStateStore;
 pub use state::{
